pyproject.toml
[tool.poetry]
name = "reldplayer"
version = "0.1.0"
description = ""
authors = ["ZackaryW <36378555+ZackaryW@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "reldplayer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
zrcl = "^0.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
black = "^24.4.2"
psutil = "^5.9.8"
pygetwindow = "^0.0.9"
pyautogui = "^0.9.54"
sphinx = "^7.3.7"
furo = "^2024.5.6"
sphinx-intl = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]